FEAT : Room DB에 데이터 저장 및 Home 화면에 데이터 보이기
+ HomeScreen에 dummy 데이터 보이기

+++ app/src/main/java/com/example/videoalarm/data/AlarmDao.kt
@@ -24,7 +24,7 @@ interface AlarmDao {
     fun getAllAlarms(): Flow<List<Alarm>>
 
     @Query("SELECT * FROM Alarm WHERE id = :id")
-    fun getAlarm(id : Int) : Flow<Alarm>
+    fun getAlarm(id : Long) : Flow<Alarm>
 
 
 }
